Abstract out duplicate search function
